Remove unused, auto-created files in specimen module
Move AmbiguousLocationException & RequestedSpecimens classes, and type methods to study-api
