--- specimen/src/org/labkey/specimen/SpecimenController.java
@@ -1,48 +1,0 @@
-/*
- * Copyright (c) 2021 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.labkey.specimen;
-
-import org.labkey.api.action.SimpleViewAction;
-import org.labkey.api.action.SpringActionController;
-import org.labkey.api.security.RequiresPermission;
-import org.labkey.api.security.permissions.ReadPermission;
-import org.labkey.api.view.JspView;
-import org.labkey.api.view.NavTree;
-import org.springframework.validation.BindException;
-import org.springframework.web.servlet.ModelAndView;
-
-public class SpecimenController extends SpringActionController
-{
-    private static final DefaultActionResolver _actionResolver = new DefaultActionResolver(SpecimenController.class);
-    public static final String NAME = "specimen";
-
-    public SpecimenController()
-    {
-        setActionResolver(_actionResolver);
-    }
-
-    @RequiresPermission(ReadPermission.class)
-    public class BeginAction extends SimpleViewAction
-    {
-        public ModelAndView getView(Object o, BindException errors)
-        {
-            return new JspView("/org/labkey/specimen/view/hello.jsp");
-        }
-
-        public void addNavTrail(NavTree root) { }
-    }
-}
