--- specimen/src/org/labkey/specimen/view/hello.jsp
@@ -1,25 +1,0 @@
-<%
-/*
- * Copyright (c) 2021 LabKey Corporation
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-%>
-<%@ page import="org.labkey.api.data.Container" %>
-<%@ page import="org.labkey.api.security.User" %>
-<%@ page extends="org.labkey.api.jsp.JspBase" %>
-<%
-    Container c = getContainer();
-    User user = getUser();
-%>
-<div name="helloMessage">Hello, and welcome to the Specimen module.</div>
