+++ study/src/org/labkey/study/query/SpecimenRequestQueryView.java
@@ -37,7 +37,6 @@ import org.labkey.api.util.PageFlowUtil;
 import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.NavTree;
 import org.labkey.api.view.ViewContext;
-import org.labkey.study.SpecimenManager;
 import org.labkey.study.controllers.specimen.SpecimenController;
 import org.labkey.study.controllers.specimen.SpecimenController.ManageRequestAction;
 import org.labkey.study.model.StudyImpl;
